refactor: move mobile web auth adapter exports
