+++ .changeset/react-native-wata-mobile-web-auth.md
@@ -2,8 +2,8 @@
 "accounts": patch
 ---
 
-Added Wata mobile web auth adapters for React Native.
+Added Wata mobile web auth adapters under `accounts/mobileWebAuth`.
 
 Added `tempoWallet()` defaults and `mobileWebAuth()` for custom Wata host, provider identity, and app-provided browser openers.
 
-Added `accounts/react-native/expo-web-browser` for Expo auth sessions.
+Added `accounts/mobileWebAuth/expoWebBrowser` for Expo auth sessions.
