+++ src/mobileWebAuth/Provider.localnet.test.ts
@@ -12,7 +12,8 @@ import * as Provider from '../core/Provider.js'
 import * as Storage from '../core/Storage.js'
 import * as Store from '../core/Store.js'
 import * as Rpc from '../core/zod/rpc.js'
-import { mobileWebAuth, tempoWallet } from './adapter.js'
+import { mobileWebAuth } from './adapter.js'
+import { tempoWallet } from './tempoWallet.js'
 
 const callback = 'xyz.tempo.accounts.playground:/auth'
 const consumerOrigin = 'https://accounts-playground.example'
