+++ src/react-native/index.ts
@@ -1,1 +1,2 @@
-export { mobileWebAuth, tempoWallet } from './adapter.js'
+export { asyncStorage } from './asyncStorage.js'
+export { secureStorage } from './secureStorage.js'
